sched.h: move pid helpers to pid.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

This is needed for killing the sched.h dependency on rcupdate.h, and
pid.h is a better place for this code anyways.

(cherry picked from commit f551103cb964e9e6f5c03b3b8723424723731e76)
